Update translations (#4)
* update text content and rspec

* Reset redirect_url query param rewriting

At the moment, the redirect_url parameter is not setted correctly.
The event "DomSubtreeModified" cause an infinite loop on Firefox. So it's recommanded to use mutation Observers.

In this sample, there is a mutationObserver Promise which should return the call to action button from modal. Then it allows to modify the redirect_url query param in the href attribute.